Address deprecation concerns
