+++ pytorch_lightning/trainer/configuration_validator.py
@@ -46,6 +46,8 @@ class ConfigValidator:
         self._check_add_get_queue(model)
         # TODO(@daniellepintz): Delete _check_progress_bar in v1.7
         self._check_progress_bar(model)
+        # TODO: Delete _check_on_post_move_to_device in v1.7
+        self._check_on_post_move_to_device(model)
         # TODO: Delete _check_on_keyboard_interrupt in v1.7
         self._check_on_keyboard_interrupt()
 
@@ -125,6 +127,19 @@ class ConfigValidator:
             rank_zero_deprecation(
                 "The `LightningModule.get_progress_bar_dict` method was deprecated in v1.5 and will be removed in v1.7."
                 " Please use the `ProgressBarBase.get_metrics` instead."
+            )
+
+    def _check_on_post_move_to_device(self, model: "pl.LightningModule") -> None:
+        r"""
+        Checks if `on_post_move_to_device` method is overriden and sends a deprecation warning.
+
+        Args:
+            model: The model to check the `on_post_move_to_device` method.
+        """  
+        if is_overridden("on_post_move_to_device", model):
+            rank_zero_deprecation(
+                "Method `on_post_move_to_device` has been deprecated in v1.5 and will be removed in v1.7. "
+                "We perform auto parameters tying without the need of implementing `on_post_move_to_device`"
             )
 
     def __verify_eval_loop_configuration(self, model: "pl.LightningModule", stage: str) -> None:
