+++ tests/deprecated_api/test_remove_1-7.py
@@ -16,6 +16,7 @@ from unittest import mock
 
 import pytest
 import torch
+from torch import nn
 
 from pytorch_lightning import Callback, LightningDataModule, Trainer
 from pytorch_lightning.loggers import LoggerCollection, TestTubeLogger
@@ -255,3 +256,30 @@ def test_v1_7_0_deprecate_lightning_distributed(tmpdir):
         from pytorch_lightning.distributed.dist import LightningDistributed
 
         _ = LightningDistributed()
+
+
+def test_v1_7_0_deprecate_on_post_move_to_device(tmpdir):
+
+    class WeightSharingModule(BoringModel):
+        def __init__(self):
+            super().__init__()
+            self.layer_1 = nn.Linear(32, 10, bias=False)
+            self.layer_2 = nn.Linear(10, 32, bias=False)
+            self.layer_3 = nn.Linear(32, 10, bias=False)
+
+        def on_post_move_to_device(self):
+            self.layer_3.weight = self.layer_1.weight
+
+        def forward(self, x):
+            x = self.layer_1(x)
+            x = self.layer_2(x)
+            x = self.layer_3(x)
+            return x
+
+    model = WeightSharingModule()
+    trainer = Trainer(default_root_dir=tmpdir, limit_train_batches=5, max_epochs=1)
+
+    with pytest.deprecated_call(
+        match=r"Method `on_post_move_to_device` has been deprecated in v1.5 and will be removed in v1.7"
+    ):
+        trainer.fit(model)
